moved lsm liquid stake test
